multiple changes, preparation for bugfixing and cron tasks
