+++ files/auftrag.php
@@ -68,6 +68,11 @@ if (isset($_GET['show'])) {
 $mitarbeiter = DBAccess::selectQuery("SELECT Vorname, Nachname, id FROM mitarbeiter");
 $colors = DBAccess::selectQuery("SELECT Farbe, Bezeichnung, Hersteller, Farbwert, id AS Nummer FROM color");
 
+$path = 'files/res/js/';
+$files = scandir($path);
+Protocoll::prettyPrint($files);
+
+
 if ($auftragsId == -1): ?>
 	<style>
 		main {
@@ -367,7 +372,7 @@ if ($auftragsId == -1): ?>
 		</form>
 		<p>Hier Dateien per Drag&Drop ablegen oder 
 			<label class="uploadWrapper">
-				<input type="file" name="uploadedFile" multiple class="fileUploadBtn" form="uploadFilesOrder">
+				<input type="file" name="uploadedFile" multiple class="fileUploadBtn" form="auftragUpload">
 				hier hochladen
 			</label>
 		</p>
