+++ files/tagsOverview.php
@@ -10,6 +10,7 @@ $query = "SELECT `content` FROM module_sticker_tags";
 $tags = DBAccess::selectQuery($query);
 
 ?>
+<script type="module" src="<?=Link::getResourcesShortLink("tags.js", "js")?>"></script>
 <p>Anzahl Tags: <?=count($tags)?></p>
 <?php foreach ($tags as $tag): ?>
     <dt class="inline-block py-1 px-2 border-none rounded-lg bg-cyan-500 m-2"><?=$tag["content"]?></dt>
